fix: check if socket.terminate exists before calling it (#2)
This PR adds a check if `socket.terminate()` function exists before calling it in the `stream.destroy` callback.

The WebSocket spec doesn't include the `terminate()` function which I suppose is what `ws` module provides. This in turn creates an "terminate is undefined" error in the browser.

Calls `socket.close` if `socket.terminate` doesn't exist.